Move SteamService and NotificationHelper to the service package
