+++ app/src/main/AndroidManifest.xml
@@ -47,7 +47,7 @@
             </intent-filter>
         </activity>
         <service
-            android:name=".SteamService"
+            android:name=".service.SteamService"
             android:enabled="true"
             android:exported="false"
             android:foregroundServiceType="dataSync" />
