+++ app/src/main/java/com/OxGames/Pluvia/PrefManager.kt
@@ -11,6 +11,7 @@ import androidx.datastore.preferences.core.intPreferencesKey
 import androidx.datastore.preferences.core.longPreferencesKey
 import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
+import com.OxGames.Pluvia.service.SteamService
 import com.OxGames.Pluvia.ui.enums.Orientation
 import com.winlator.box86_64.Box86_64Preset
 import com.winlator.container.Container
