+++ app/src/main/java/com/OxGames/Pluvia/service/NotificationHelper.kt
@@ -1,4 +1,4 @@
-package com.OxGames.Pluvia
+package com.OxGames.Pluvia.service
 
 import android.app.Notification
 import android.app.NotificationChannel
@@ -9,6 +9,8 @@ import android.content.Context.NOTIFICATION_SERVICE
 import android.content.Intent
 import androidx.core.app.NotificationCompat
 import androidx.core.net.toUri
+import com.OxGames.Pluvia.MainActivity
+import com.OxGames.Pluvia.R
 
 class NotificationHelper(private val context: Context) {
