+++ app/src/main/java/com/OxGames/Pluvia/ui/model/MainViewModel.kt
@@ -6,12 +6,12 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.OxGames.Pluvia.PluviaApp
 import com.OxGames.Pluvia.PrefManager
-import com.OxGames.Pluvia.SteamService
 import com.OxGames.Pluvia.data.GameProcessInfo
 import com.OxGames.Pluvia.enums.LoginResult
 import com.OxGames.Pluvia.enums.PathType
 import com.OxGames.Pluvia.events.AndroidEvent
 import com.OxGames.Pluvia.events.SteamEvent
+import com.OxGames.Pluvia.service.SteamService
 import com.OxGames.Pluvia.ui.data.MainState
 import com.OxGames.Pluvia.ui.enums.PluviaScreen
 import com.winlator.xserver.Window
